Refactor RelayCommand location
